dotnet reference documentation and tweaks to the output
