+++ docs/reference/dotnet/Google_Protobuf_WellKnownTypes.md
@@ -1,0 +1,40 @@
+#### [Trinsic](index.md 'index')
+## Google.Protobuf.WellKnownTypes Namespace
+### Classes
+<a name='Google_Protobuf_WellKnownTypes_StructExtensions'></a>
+## StructExtensions Class
+```csharp
+public static class StructExtensions
+```
+
+Inheritance [System.Object](https://docs.microsoft.com/en-us/dotnet/api/System.Object 'System.Object') &#129106; StructExtensions  
+### Methods
+<a name='Google_Protobuf_WellKnownTypes_StructExtensions_ToJObject(Google_Protobuf_WellKnownTypes_Struct)'></a>
+## StructExtensions.ToJObject(Struct) Method
+Converts to `JObject`.  
+```csharp
+public static Newtonsoft.Json.Linq.JObject ToJObject(this Google.Protobuf.WellKnownTypes.Struct message);
+```
+#### Parameters
+<a name='Google_Protobuf_WellKnownTypes_StructExtensions_ToJObject(Google_Protobuf_WellKnownTypes_Struct)_message'></a>
+`message` [Google.Protobuf.WellKnownTypes.Struct](https://docs.microsoft.com/en-us/dotnet/api/Google.Protobuf.WellKnownTypes.Struct 'Google.Protobuf.WellKnownTypes.Struct')  
+The message.
+  
+#### Returns
+[Newtonsoft.Json.Linq.JObject](https://docs.microsoft.com/en-us/dotnet/api/Newtonsoft.Json.Linq.JObject 'Newtonsoft.Json.Linq.JObject')  
+  
+<a name='Google_Protobuf_WellKnownTypes_StructExtensions_ToJToken(Google_Protobuf_WellKnownTypes_Struct)'></a>
+## StructExtensions.ToJToken(Struct) Method
+Converts to jtoken.  
+```csharp
+public static Newtonsoft.Json.Linq.JToken ToJToken(this Google.Protobuf.WellKnownTypes.Struct message);
+```
+#### Parameters
+<a name='Google_Protobuf_WellKnownTypes_StructExtensions_ToJToken(Google_Protobuf_WellKnownTypes_Struct)_message'></a>
+`message` [Google.Protobuf.WellKnownTypes.Struct](https://docs.microsoft.com/en-us/dotnet/api/Google.Protobuf.WellKnownTypes.Struct 'Google.Protobuf.WellKnownTypes.Struct')  
+The message.
+  
+#### Returns
+[Newtonsoft.Json.Linq.JToken](https://docs.microsoft.com/en-us/dotnet/api/Newtonsoft.Json.Linq.JToken 'Newtonsoft.Json.Linq.JToken')  
+  
+
