+++ docs/reference/dotnet/Newtonsoft_Json_Linq.md
@@ -1,0 +1,25 @@
+#### [Trinsic](index.md 'index')
+## Newtonsoft.Json.Linq Namespace
+### Classes
+<a name='Newtonsoft_Json_Linq_JTokenExtensions'></a>
+## JTokenExtensions Class
+```csharp
+public static class JTokenExtensions
+```
+
+Inheritance [System.Object](https://docs.microsoft.com/en-us/dotnet/api/System.Object 'System.Object') &#129106; JTokenExtensions  
+### Methods
+<a name='Newtonsoft_Json_Linq_JTokenExtensions_ToStruct(Newtonsoft_Json_Linq_JToken)'></a>
+## JTokenExtensions.ToStruct(JToken) Method
+Converts to struct.  
+```csharp
+public static Google.Protobuf.WellKnownTypes.Struct ToStruct(this Newtonsoft.Json.Linq.JToken token);
+```
+#### Parameters
+<a name='Newtonsoft_Json_Linq_JTokenExtensions_ToStruct(Newtonsoft_Json_Linq_JToken)_token'></a>
+`token` [Newtonsoft.Json.Linq.JToken](https://docs.microsoft.com/en-us/dotnet/api/Newtonsoft.Json.Linq.JToken 'Newtonsoft.Json.Linq.JToken')  
+  
+#### Returns
+[Google.Protobuf.WellKnownTypes.Struct](https://docs.microsoft.com/en-us/dotnet/api/Google.Protobuf.WellKnownTypes.Struct 'Google.Protobuf.WellKnownTypes.Struct')  
+  
+
