+++ docs/reference/dotnet/index.md
@@ -1,0 +1,17 @@
+#### [Trinsic](index.md 'index')
+
+| Namespaces | |
+| :--- | :--- |
+| [Google.Protobuf](Google_Protobuf.md 'Google.Protobuf') |  |
+| [Google.Protobuf.WellKnownTypes](Google_Protobuf_WellKnownTypes.md 'Google.Protobuf.WellKnownTypes') |  |
+| [Newtonsoft.Json.Linq](Newtonsoft_Json_Linq.md 'Newtonsoft.Json.Linq') |  |
+| [Trinsic](Trinsic.md 'Trinsic') |  |
+| [Trinsic.Sdk.Options.V1](Trinsic_Sdk_Options_V1.md 'Trinsic.Sdk.Options.V1') |  |
+| [Trinsic.Services.Account.V1](Trinsic_Services_Account_V1.md 'Trinsic.Services.Account.V1') |  |
+| [Trinsic.Services.Common.V1](Trinsic_Services_Common_V1.md 'Trinsic.Services.Common.V1') |  |
+| [Trinsic.Services.Debug.V1](Trinsic_Services_Debug_V1.md 'Trinsic.Services.Debug.V1') |  |
+| [Trinsic.Services.Provider.V1](Trinsic_Services_Provider_V1.md 'Trinsic.Services.Provider.V1') |  |
+| [Trinsic.Services.TrustRegistry.V1](Trinsic_Services_TrustRegistry_V1.md 'Trinsic.Services.TrustRegistry.V1') |  |
+| [Trinsic.Services.UniversalWallet.V1](Trinsic_Services_UniversalWallet_V1.md 'Trinsic.Services.UniversalWallet.V1') |  |
+| [Trinsic.Services.VerifiableCredentials.Templates.V1](Trinsic_Services_VerifiableCredentials_Templates_V1.md 'Trinsic.Services.VerifiableCredentials.Templates.V1') |  |
+| [Trinsic.Services.VerifiableCredentials.V1](Trinsic_Services_VerifiableCredentials_V1.md 'Trinsic.Services.VerifiableCredentials.V1') |  |
