+++ dotnet/Trinsic/Trinsic.csproj
@@ -4,6 +4,8 @@
         <TargetFrameworks>net6.0;net6.0-ios</TargetFrameworks>
         <Nullable>enable</Nullable>
         <IsBrowser>false</IsBrowser>
+        <GenerateDocumentationFile>true</GenerateDocumentationFile>
+        <DocumentationFile>Trinsic.xml</DocumentationFile>
     </PropertyGroup>
 
     <PropertyGroup Condition="'$(IsBrowser)' == 'true'">
@@ -13,7 +15,7 @@
     <ItemGroup Condition="'$(IsBrowser)' == 'true'">
         <PackageReference Include="Grpc.Net.Client.Web" Version="2.40.0" />
     </ItemGroup>
-
+    
     <!--
       https://github.com/novotnyllc/MSBuildSdkExtras/issues/176
       https://githubmemory.com/repo/xamarin/xamarin-macios/issues/12631
